apps: pkt_echo: added port loopback mode when working with 2 ports.
This mode is supported when working with 2 or 4 cores.

In case of 2 cores: one core will work with one port and
second with the other. Only queue 0 will be used (probably
RSS configuration is needed).

In case of 4 cores: 2 cores will work with one port and
2 other cores with the other. Queues 0 and 1 will be used.
RSS configuration is required.

Change-Id: Id8ebcb768977e6e10c4549f21fb72a5f7860db29
Signed-off-by: nsamsono <nsamsono@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/36059
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Yuval Caduri <cyuval@marvell.com>