sam: kat: Add optional "Textlen" parameter
If "Textlen" doesn't exist in the input file or 0 -
no change in current behavior.
Textlen is derived from Plaintext/Ciphertext string

If Textlen is exist and not 0 in the input file,
Plaintext and Ciphertext will became of Textlen size.
If Plaintext and Ciphertext strings in the input file
are large than Textlen, they will be cut to Textlen size.
If Plainitext and Ciphertext strings in the input file
are shorter than Textlen, they will be extended to Textlen
size by multiple usage of original Plaintext/Ciphertext strings.

File aes_1440_multi.txt has three performance tests now
- AES-CBS 		- Rate: 484 Kpps, 5575 Mbps
- AES-CBC + SHA1 (new)	- Rate: 434 Kpps, 4999 Mbps
- AES-CBC + MD5  (new)	- Rate: 300 Kpps, 3456 Mbps

Change-Id: I8baa54d92e921596eae9d7430d74451b728dae98
Signed-off-by: Dmitri Epshtein <dima@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/39945
Tested-by: iSoC Platform CI <ykjenk@marvell.com>