fix: dmax2: avoid exposing DMAX2 struct to user application
The main DMA structure struct dmax2 was returned as is to registering
application. This can lead to issues, since this way the app gains
access the private DMAX2 fields.

This patch fixes this issue by returning the DMAX2 struct as a handle
(e.g. void *) so it's fields are not be exposed to user app.

Change-Id: Icc015b486e1ffbdfbd3a6b5e7c79367642a5d53e
Signed-off-by: Omri Itach <omrii@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/41267
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Yehuda Yitschak <yehuday@marvell.com>