fix: dmax2: don't fail enqueue operation in case not enough queue space
The DMA driver didn't allow application to call enqueue operation
with more descriptors than available.
In case app tried to enqueue more than available, it returned an error
and didn't enqueue anything.

This is now fixed to be a warning, since it's a valid standard scenario
and should be handled as a warning, even though application can avoid
this state by checking how much descriptors are available to be queued
before operation.

Change-Id: Id0109fe8ee930109814ebf09f2fcc2030a504b9c
Signed-off-by: Omri Itach <omrii@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/43604
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Yehuda Yitschak <yehuday@marvell.com>