gie: fix head/tail idx override during dma copy
when we call gie_copy_idx we instruct the DMA to copy the value pointed
by q->head and q->tail. If before this DMA transaction completes,
another batch is processed, then the value of local shadows q->head/tail
is modified before the DMA operation completed.
This causes an issue since the shadow indices are incremented before the
QEs and buffers are actaully copied.
To resolve this, we create a queue that holds the values of the indices
to be copied. This way we can always modify the shadows since the value
to copy is stored in a backup ring.

Change-Id: Ief963dcef46e8829e8a57f5d7627eff4d3cefdec
Signed-off-by: Yehuda Yitschak <yehuday@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/45047
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Shadi Ammouri <shadi@marvell.com>