gie: fix remote queue QEs copy override
In egress path (i.e. remote queue handling), QEs are copied
using DMA to destination queue. The index to which these QEs are copied
is controlled by qe_tail.
This patch limits the number of QEs to be copied according to the
consumer pointer of the destination queue (i.e. the head index).
Added a new macro ("qes_copy_space") to calculate the empty space in the
destnination queue that new QEs can be copied into, and use this macro
to clip the number of QEs to be copied.

Change-Id: I7a4c7ba9b1ce1b6d6980bd59ceffbdac86f6b8e8
Signed-off-by: Shadi Ammouri <shadi@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/46240
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Yehuda Yitschak <yehuday@marvell.com>