giu: gpio: fix: manage shadow-q of indexes
problem: on RSS mode, as a result of wrong indication on the number of
transmitted frames, buffers of un-transmitted frames were released to
the pp2-pool which override the data and cause a mismatch in netdev
side.

fix: created a shadow-q with the same size as tx-queue which will holds
for each sent frame its 'tx-queue-index' (as a result of RSS) and its
entry index in this tx-queue.
once application request for the number of transmitted frames, we go
over on all occupied entries in the shadow-q and using the saved
information check if it was transmitted or not. once we reach
an un-transmitted frame we stop and return the number of transmitted
frames. as the shadow-q holds the frames in the same order as the
application sent them, it is guarantee that only transmitted buffers
will be released to the pp2-pool.

FIX AXEL-575

Change-Id: Ifa7315ef227fa1c0924f40c2abce9b02f51f2d1d
Signed-off-by: Liron Himi <lironh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/50964
Reviewed-by: Alan Winkowski <walan@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Yuval Caduri <cyuval@marvell.com>
(cherry picked from commit f33366e1c9aefba5386ee59222a24dbe29c90a2c)
Reviewed-on: http://vgitil04.il.marvell.com:8080/51017