msi: update and register msix table base
GIE: defined new API for MSI-X table registration

NMP: The MSI-X table offset is updated in PCI configuration
area. In addition, it is registered to GIE.

Note that the table address it cannot be set at init time as the GIE
is initialized before the PCI configuration space is
updated (by the host and the device).

Change-Id: Ie625c41791aee3bf1c4d4d5614bbe41da226ecf7
Signed-off-by: Haim Boot <hayim@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/53002
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Liron Himi <lironh@marvell.com>