giu: read queue indexes from real location
Relaunching guest app where in the first time traffic was involves
may cause some indexes issues.
This is because we are using shadow indexes which assumes it start
with ‘0’ but in the 2nd it may point to other index.

this patch remove those 'shadow' indexes and read the values
from the real location

Change-Id: I2865ae0fe50438b38a8191eca122192eb9b8acad
Signed-off-by: Liron Himi <lironh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/53524
Tested-by: iSoC Platform CI <ykjenk@marvell.com>