giu: gpio: optimize RSS calculation
Currently in order to calculate the hash value of a key, it is required
to extract the fields and build a new key (using memcpy).
This change provide the ability to calculate the hash without the need
for memcpy and just pass the required fields.
The hash function was modified to get previous hash result and
accumulate to it the input data.

Change-Id: I77a780d6578583eda596ff1be0b2971fb1e2ab5d
Signed-off-by: Liron Himi <lironh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/55321
Reviewed-by: Alan Winkowski <walan@marvell.com>