mng: fix: allocate msix tables in bar0 range
BAR0 size was 1MB.
The MSI-X tables were allocated outside this range which
caused them to be overrun by other drivers which allocated
this area for their use.

This patch set the MSI-X tables inside BAR0 range.
In addition, it adds a check that the allocated size of
BAR0 is big enough to contain its regions.

Change-Id: I27be87dac1e25e3b30fcc090699bc2c1df0ffd50
Signed-off-by: Haim Boot <hayim@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/58057
Reviewed-by: Shlomi Gridish <sgridish@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Liron Himi <lironh@marvell.com>