pf: stat: add gp_rx_fullq_dropped counter
This new counter relies on pp2 rx counter so
the pp2 statistics are now saved in the port descriptor
so they can be accessed by pf level as well (and
not just by pf_pp2).

when gp statistics are read, the pp2 statistics are
taken from the port descriptor and gp_rx_fullq_dropped
counter is calculated using pp2 rx counter and gp
rx counter.

Change-Id: Icdb66e53bcdcb9c3ada09d99da6e85035a1ad84d
Signed-off-by: Haim Boot <hayim@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/58285
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Liron Himi <lironh@marvell.com>