apps: pp2_utils: fix: read packets according to available space in shadow queue
In the previous implementation, we first read the number of requested
packets and only then calculated the available space in the shadow queue.
in case there was not enough space, it dropped the rest of the packets.

In this patch, we first calculate the available space, update the number
of packets to read (in case there is not enough space) and only then
read the packets.

Change-Id: I95dba6eb92fb2555e0c36355cfc220f48f66c696
Signed-off-by: Haim Boot <hayim@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/59010
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Alan Winkowski <walan@marvell.com>
(cherry picked from commit decadf740c0076132cd640dbfd5c4c288c744aa0)
Reviewed-on: http://vgitil04.il.marvell.com:8080/59183