neta: fix: flush ucast only in non-promisc mode
in case user changed the mac-address while the port was in
promisc mode, the current code reset the ucast HW regs which blocked all
unicast frames; i.e. port is not in promisc mode anymore

the fix checks if the port is in promisc and avoid the reset of the
hw regs. those regs will be populated with the SW DB once the port
exist from the promisc mode.

fix ODP-4

Change-Id: I04a3c1f38241265bda49f178f144f45abd163751
Signed-off-by: Liron Himi <lironh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/61911
Reviewed-by: Alan Winkowski <walan@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/5534
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Liron Himi <liron.himi@marvell.com>