apps: neta: pkt-echo: use common function for cpu mask creation
Use apps_cores_mask_create() function for creating the
cpu mask for the application. This function is used in
pp2 pkt-echo also.
The advantage of using this function is that it starts
from cpu_mask = 0x2 (and not 0x1) in case there are
2 cpus available (meaning the application will run
on the second core and not the first core).

Change-Id: I34febdebcde7408fddde58c7db8b91fea949c138
Signed-off-by: Alan Winkowski <walan@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/6763
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Yuri Chipchev <yuri.chipchev@marvell.com>
Reviewed-by: Liron Himi <liron.himi@marvell.com>