giu: gie: handle s/g frames
The current assumptions is that both local and remote buffers have
the same size and there is only one local pool.
with that assumptions, GIE just need to check there are enough
buffers to handle a complete s/g frame, if not it stopped processing
and the same frame will be handled next time.

Change-Id: I72718430d9d5d85e653ddf353790ba95912e4fa3
Signed-off-by: Liron Himi <lironh@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/22759
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Yuri Chipchev <yuric@marvell.com>