pp2: ppio: fix tx queue counter
there are 2 counters per queue, one for SW enqueue
and the other for HW dequeue.
usually they are the same, except where tx port
is stuck and doesn't dequeue from the HW queue.
in this case, the enqueue will increment by SW, but
the dequeue sill not.
it is more accurate to read the dequeue counter
as the transmitted packets

Change-Id: I6b7ebe7294d54819b86eae183e1ac94f8fbff058
Signed-off-by: Liron Himi <lironh@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/musdk/+/37808
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>